requirements.txt
streamlit >= 0.79.0
plotly-express
umap
pillow
pandas >= 1.0.5
Image >=1.5.33
stmol